requirements.txt
slackclient
chalice
boto3
praw
requests